refactor: rename table.ts with api-requests
